Version Translate 2.3
Board is a little larger and it incorporates both connectors which are used for driver and sensor board on Rev2 controler. I recommend this board if you decide to go that route.
Additionally i included the inverter chip connected to BOTH driver connectors sou you can drive both IGBT bridges if you want to experiment.